test(htmx): consolidate browser-e2e harness + drop stale settings.js shim
Two follow-ups from PR 131's second code-review pass plus the
"keep browser-e2e post-#129" decision:

1. **Delete the stale `data-ryokan-confirm-label` shim in
   `static/js/settings.js`.** The IIFE at the top of the file was a
   near-duplicate of `base.js`'s `ryokanConfirmFromAttrs` shim that
   still read the now-renamed `data-ryokan-confirm-label` attribute.
   Net runtime effect was a doubled modal call per submit (settings.js
   shim opened a modal labeled "Yes," base.js's listener immediately
   overwrote it with the right label) plus an unresolved Promise
   leaking per click. The CF-Yes-button regression added in PR 131's
   first fix-up still passes by accident — both calls happen
   synchronously within the same submit handler, so the wrong-label
   modal is invisible. Single source of truth now lives in base.js
   (DOMContentLoaded handler for native form-POST forms +
   `htmx:confirm` body listener for HTMX forms).

2. **Consolidate browser-e2e harness boilerplate** to
   `tests/common/browser_e2e.rs` (per PR 131 review item 8). Each
   test binary pulls it via `#[path = "common/browser_e2e.rs"]
   mod browser_e2e;`. Cargo skips files under `tests/common/` for
   test-binary discovery so it never produces a "0 tests passed"
   line. Net change: ~860 LoC removed from the three e2e files,
   ~540 LoC added to the new shared module. Module carries
   `#![allow(dead_code)]` because each binary uses a subset of the
   helpers — every binary compiles the full module + emits dead-code
   warnings for the helpers it doesn't happen to call; "unused in
   this binary" isn't actionable signal for shared infrastructure.

   Helpers consolidated: `spawn_app`, `try_connect_browser`,
   `resolve_browser_binary`, `librewolf_shim`, `shell_quote`, `which`,
   `seed_user_session`, `open_with_session`, `assert_htmx_loaded`,
   `assert_htmx_handled_in_place`, `assert_modal_text`,
   `assert_dom_contains`, `wait_for_row_removed`,
   `wait_for_confirm_modal`, `wait_until_substring`,
   `click_delete_for`. Module docstring captures the per-helper
   purpose + the four conventions for new browser-e2e tests
   (assert_htmx_handled_in_place + survivor-row check + DB-side
   side-effect verification + modal-copy assertion) that emerged
   from the false-positive audit.

3. **Reframe browser-e2e as general-purpose** rather than
   "scheduled for removal post-#129." The harness has paid for itself
   already (caught 4 false positives during the Phase 1.5 audit + the
   CF Yes-label bug from the PR 131 review) and the marginal cost to
   keep it is low (gated behind `--features browser-e2e`; not in CI
   default; no prod-build impact). Future template-driven UI work
   (Phase 7 frontend rework, #125 bulk library ops, #121 notification
   settings, #116 calendar, etc.) inherits the harness as-is. Updated
   CLAUDE.md accordingly. Test files no longer reference removal.

All 10 browser e2e tests still green vs LibreWolf. 1951 unit + 47
integration tests still green. fmt + clippy clean on both feature
shapes.
