+++ tests/htmx_browser_e2e.rs
@@ -28,11 +28,10 @@
 //!
 //! Tests are NOT enabled in CI: they require an out-of-band driver +
 //! browser binary, and the goal is local-iteration speed during the
-//! HTMX migration, not gating CI on browser availability. The whole
-//! `browser-e2e` feature (Cargo.toml flag, fantoccini dev-dep, this
-//! file, and the `e2e_browser_app` helper in `test_support.rs`) is
-//! planned to be removed once issue #129 is closed — search for
-//! `browser-e2e` to find every removal site.
+//! HTMX migration. Browser e2e is now considered general-purpose
+//! infrastructure for future template-driven UI work (#125 bulk ops,
+//! #121 notification settings, #116 calendar, Phase 7 rework, etc.) —
+//! it is no longer scheduled for removal post-#129.
 //!
 //! ## When the test gracefully skips
 //!
@@ -44,170 +43,14 @@
 //! laptop without geckodriver pre-started, which matches how CI-gated
 //! `live_smoke` tests (`RYOKAN_QBIT_E2E` etc.) opt into a real daemon.
 
-use std::net::SocketAddr;
 use std::time::Duration;
 
-use fantoccini::ClientBuilder;
 use fantoccini::Locator;
-use ryokan::test_support::{
-    build_test_app_state, e2e_browser_app, in_memory_pool, logged_in_session,
-};
+use ryokan::test_support::{build_test_app_state, in_memory_pool, logged_in_session};
 
-/// Spawn the e2e browser app on a random local port; return the bound
-/// address and a JoinHandle that owns the axum task. The handle is
-/// dropped at end of test, which aborts the task — no graceful
-/// shutdown plumbing because tests should not depend on it.
-async fn spawn_app(state: ryokan::AppState) -> SocketAddr {
-    let listener = tokio::net::TcpListener::bind("127.0.0.1:0")
-        .await
-        .expect("bind random port");
-    let addr = listener.local_addr().expect("local_addr");
-    let app = e2e_browser_app(state);
-    tokio::spawn(async move {
-        axum::serve(listener, app).await.expect("serve");
-    });
-    // Tiny settle so the listener is in `accept()` before the browser
-    // dials it. axum::serve returns from bind synchronously, but a
-    // first connect-immediately-after-bind on some kernels eats the SYN.
-    tokio::time::sleep(Duration::from_millis(50)).await;
-    addr
-}
-
-/// Connect to the WebDriver endpoint, returning Ok(client) on success
-/// or Err(message) when the connect fails for any reason. Tests use
-/// the Err path to print a graceful skip message and return.
-///
-/// Browser binary resolution: geckodriver defaults to looking for
-/// `firefox` in PATH, which fails on a LibreWolf-only system (and on
-/// CI-like images that have a non-default install path). The harness
-/// honors `RYOKAN_BROWSER_BIN` as an explicit override; absent that,
-/// it auto-detects `firefox` / `firefox-esr` / `librewolf` in PATH
-/// and passes the first hit through `moz:firefoxOptions.binary`.
-///
-/// **LibreWolf shim**: geckodriver runs `<binary> --version` and only
-/// accepts output that starts with `Mozilla Firefox`. LibreWolf prints
-/// `Mozilla LibreWolf X.Y`, which trips a "binary is not a Firefox
-/// executable" hard-fail before any test code runs. When the resolved
-/// binary is LibreWolf, the harness writes a tiny wrapper script to a
-/// tempfile that fakes the `--version` line and execs LibreWolf for
-/// every other invocation, then hands geckodriver the wrapper path.
-/// Survives until the next `tmp` cleanup; not worth a Drop guard.
-async fn try_connect_browser() -> Result<fantoccini::Client, String> {
-    let url = std::env::var("RYOKAN_WEBDRIVER_URL")
-        .unwrap_or_else(|_| "http://localhost:4444".to_string());
-
-    let mut caps = serde_json::Map::new();
-    // Headless by default — devs running locally can flip with
-    // RYOKAN_BROWSER_HEADLESS=0 if they want to watch the test drive
-    // the browser visually for debugging. CI-shape (no driver running)
-    // never reaches this code path.
-    let headless = std::env::var("RYOKAN_BROWSER_HEADLESS")
-        .map(|v| v != "0" && !v.eq_ignore_ascii_case("false"))
-        .unwrap_or(true);
-    let mut firefox_opts = serde_json::Map::new();
-    if headless {
-        firefox_opts.insert("args".to_string(), serde_json::json!(["-headless"]));
-    }
-    if let Some(bin) = resolve_browser_binary() {
-        firefox_opts.insert("binary".to_string(), serde_json::json!(bin));
-    }
-    caps.insert(
-        "moz:firefoxOptions".to_string(),
-        serde_json::Value::Object(firefox_opts),
-    );
-
-    ClientBuilder::native()
-        .capabilities(caps)
-        .connect(&url)
-        .await
-        .map_err(|e| format!("WebDriver at {url} unavailable: {e}"))
-}
-
-fn resolve_browser_binary() -> Option<String> {
-    if let Ok(explicit) = std::env::var("RYOKAN_BROWSER_BIN")
-        && !explicit.is_empty()
-    {
-        // Explicit override is taken at face value — caller knows what
-        // they're doing. No version-shim wrapping.
-        return Some(explicit);
-    }
-    // Real Firefox is the happy path; check it before LibreWolf so a
-    // dual-install system doesn't pay the shim cost.
-    for candidate in ["firefox", "firefox-esr"] {
-        if let Some(path) = which(candidate) {
-            return Some(path);
-        }
-    }
-    if let Some(librewolf) = which("librewolf") {
-        return Some(librewolf_shim(&librewolf));
-    }
-    None
-}
-
-/// Write a wrapper script to a tempfile that intercepts `--version`
-/// (returning a Firefox-shaped string geckodriver accepts) and execs
-/// LibreWolf for every other call. Returns the wrapper path.
-///
-/// We can't just symlink — geckodriver follows the link and runs the
-/// real binary's `--version`. The wrapper has to stand between
-/// geckodriver and LibreWolf for that one call.
-fn librewolf_shim(librewolf_path: &str) -> String {
-    let wrapper_dir = std::env::temp_dir().join("ryokan-librewolf-shim");
-    std::fs::create_dir_all(&wrapper_dir).expect("create shim dir");
-    // PID in the filename so the three e2e binaries (when `cargo test`
-    // runs them in parallel) each get their own wrapper. Identical
-    // content otherwise — the race was on torn reads of the file body
-    // mid-rewrite, not on the path itself.
-    let wrapper_path = wrapper_dir.join(format!("firefox-shim-{}.sh", std::process::id()));
-    let body = format!(
-        "#!/bin/sh\n\
-         if [ \"$1\" = \"--version\" ] || [ \"$1\" = \"-version\" ]; then\n\
-         \techo \"Mozilla Firefox 149.0\"\n\
-         \texit 0\n\
-         fi\n\
-         exec {} \"$@\"\n",
-        shell_quote(librewolf_path),
-    );
-    std::fs::write(&wrapper_path, body).expect("write shim");
-    #[cfg(unix)]
-    {
-        use std::os::unix::fs::PermissionsExt;
-        let mut perms = std::fs::metadata(&wrapper_path)
-            .expect("stat shim")
-            .permissions();
-        perms.set_mode(0o755);
-        std::fs::set_permissions(&wrapper_path, perms).expect("chmod shim");
-    }
-    wrapper_path.to_string_lossy().into_owned()
-}
-
-fn shell_quote(s: &str) -> String {
-    // POSIX-safe single-quote wrap; embedded single-quotes are escaped
-    // by the standard `'\''` close-escape-open dance. Good enough for
-    // any path on a normal Linux filesystem.
-    let mut out = String::with_capacity(s.len() + 2);
-    out.push('\'');
-    for ch in s.chars() {
-        if ch == '\'' {
-            out.push_str("'\\''");
-        } else {
-            out.push(ch);
-        }
-    }
-    out.push('\'');
-    out
-}
-
-fn which(bin: &str) -> Option<String> {
-    let path = std::env::var_os("PATH")?;
-    for dir in std::env::split_paths(&path) {
-        let candidate = dir.join(bin);
-        if candidate.is_file() {
-            return Some(candidate.to_string_lossy().into_owned());
-        }
-    }
-    None
-}
+#[path = "common/browser_e2e.rs"]
+mod browser_e2e;
+use browser_e2e::{spawn_app, try_connect_browser};
 
 #[tokio::test]
 async fn episode_monitor_button_swaps_in_browser() {
@@ -225,8 +68,6 @@ async fn episode_monitor_button_swaps_in_browser() {
     // then preload the cookie via WebDriver before navigating, so the
     // first page load already passes `require_auth`.
     let (_state2, cookie_value) = logged_in_session(&db).await;
-    // cookie_value is "session=<hex>"; split off the value for the
-    // WebDriver Cookie payload (it expects name + value separately).
     let token = cookie_value
         .strip_prefix("session=")
         .expect("cookie helper returns session=<hex>")
@@ -245,15 +86,73 @@ async fn episode_monitor_button_swaps_in_browser() {
     // Drive the test, but always close the browser at the end — even
     // on assertion panics — so a stuck session doesn't leave a hung
     // browser process around.
-    let result = run_episode_monitor_test(&client, addr, series_id, &token).await;
+    let result = async {
+        // WebDriver requires a page in the target domain to be loaded
+        // before `add_cookie` will accept a cookie scoped to it. Land
+        // on /login first (always serves; no auth) and drop the
+        // session cookie in.
+        let base = format!("http://{addr}");
+        client.goto(&format!("{base}/login")).await?;
+        let raw = format!("session={token}; Path=/; SameSite=Lax");
+        let cookie = fantoccini::cookies::Cookie::parse(raw)?;
+        client.add_cookie(cookie).await?;
+
+        // Now load the test fixture page that renders the partial.
+        client
+            .goto(&format!(
+                "{base}/__test/episode-monitor-fixture?series_id={series_id}&episode_number=1"
+            ))
+            .await?;
+
+        // Pre-click assertion: button reads "Yes" because we seeded
+        // `monitored = true`.
+        let button = client.find(Locator::Css("button.ep-mon-btn")).await?;
+        let pre_text = button.text().await?;
+        assert_eq!(
+            pre_text.trim(),
+            "Yes",
+            "fixture should render the seeded monitored=true state"
+        );
+        let pre_class = button.attr("class").await?.unwrap_or_default();
+        assert!(
+            pre_class.contains("ep-mon-yes"),
+            "pre-click button should carry ep-mon-yes class; got `{pre_class}`"
+        );
+
+        // Click. htmx fires the POST, the handler returns a fresh
+        // button partial with `monitored = false`, and
+        // `hx-swap=outerHTML` replaces the element. WebDriver hands
+        // back stale-element after a swap — re-query rather than
+        // reusing the old handle.
+        button.click().await?;
+
+        // Wait for the swap to land.
+        let deadline = std::time::Instant::now() + Duration::from_secs(3);
+        loop {
+            let btn = client.find(Locator::Css("button.ep-mon-btn")).await?;
+            let text = btn.text().await?.trim().to_string();
+            let class = btn.attr("class").await?.unwrap_or_default();
+            if text == "No" && class.contains("ep-mon-no") {
+                break;
+            }
+            if std::time::Instant::now() > deadline {
+                panic!(
+                    "button did not swap to monitored=false within 3s; \
+                     last text={text:?}, class={class:?}"
+                );
+            }
+            tokio::time::sleep(Duration::from_millis(50)).await;
+        }
+        Ok::<(), Box<dyn std::error::Error>>(())
+    }
+    .await;
     let _ = client.close().await;
     result.expect("episode monitor browser test");
 
     // Side-effect verification: the handler renders the partial from
     // the request's `monitored` value, so a no-op handler that skips
     // the DB write would still return a "monitored=false" button and
-    // the browser test above would pass. Read the row back to
-    // confirm the flip actually persisted.
+    // the browser test above would pass. Read the row back to confirm.
     let states = ryokan::models::monitoring::get_series_states(&db, series_id)
         .await
         .expect("read episode states");
@@ -263,75 +162,4 @@ async fn episode_monitor_button_swaps_in_browser() {
         Some(false),
         "DB row must reflect the post-click state; got {ep1:?}"
     );
-}
-
-async fn run_episode_monitor_test(
-    client: &fantoccini::Client,
-    addr: SocketAddr,
-    series_id: i64,
-    session_token: &str,
-) -> Result<(), Box<dyn std::error::Error>> {
-    // WebDriver requires a page in the target domain to be loaded
-    // before `add_cookie` will accept a cookie scoped to it. Land on
-    // the /login page first (unauthenticated route, always serves)
-    // and then drop the session cookie in.
-    let base = format!("http://{addr}");
-    client.goto(&format!("{base}/login")).await?;
-    // Production sets SameSite=Lax. fantoccini's default cookie comes
-    // with SameSite=None (no Secure), which modern browsers reject as
-    // an unsafe combination. fantoccini re-exports `cookie::Cookie`
-    // but NOT the `SameSite` enum, so the path of least resistance is
-    // to parse a Set-Cookie-shaped string into the cookie — that route
-    // doesn't need access to the enum at all.
-    let raw = format!("session={session_token}; Path=/; SameSite=Lax");
-    let cookie = fantoccini::cookies::Cookie::parse(raw).expect("parse cookie");
-    client.add_cookie(cookie).await?;
-
-    // Now load the test fixture page that renders the partial.
-    client
-        .goto(&format!(
-            "{base}/__test/episode-monitor-fixture?series_id={series_id}&episode_number=1"
-        ))
-        .await?;
-
-    // Pre-click assertion: button reads "Yes" because we seeded
-    // `monitored = true`.
-    let button = client.find(Locator::Css("button.ep-mon-btn")).await?;
-    let pre_text = button.text().await?;
-    assert_eq!(
-        pre_text.trim(),
-        "Yes",
-        "fixture should render the seeded monitored=true state"
-    );
-    let pre_class = button.attr("class").await?.unwrap_or_default();
-    assert!(
-        pre_class.contains("ep-mon-yes"),
-        "pre-click button should carry ep-mon-yes class; got `{pre_class}`"
-    );
-
-    // Click. htmx fires the POST, the handler returns a fresh button
-    // partial with `monitored = false`, and `hx-swap=outerHTML`
-    // replaces the element. WebDriver hands back stale-element after
-    // a swap — re-query rather than reusing the old handle.
-    button.click().await?;
-
-    // Wait for the swap to land. htmx ajax + render is sub-100ms
-    // locally; poll up to ~3s for CI/slow-laptop headroom.
-    let deadline = std::time::Instant::now() + Duration::from_secs(3);
-    loop {
-        let btn = client.find(Locator::Css("button.ep-mon-btn")).await?;
-        let text = btn.text().await?.trim().to_string();
-        let class = btn.attr("class").await?.unwrap_or_default();
-        if text == "No" && class.contains("ep-mon-no") {
-            break;
-        }
-        if std::time::Instant::now() > deadline {
-            panic!(
-                "button did not swap to monitored=false within 3s; last text={text:?}, class={class:?}"
-            );
-        }
-        tokio::time::sleep(Duration::from_millis(50)).await;
-    }
-
-    Ok(())
 }
